fixed call edit ingrediente

+++ backend/app.js
@@ -500,12 +500,16 @@ app.put('/api/ingredientes/:id', async (req, res) => {
     const updateIngredientesQuery = 'UPDATE ingredientes SET nombre = $1, unidad = $2, precio = $3, proveedor = $4, proveedor_id = $5, store_route_order = $7, producto_clave = $8, moral_demanda_semanal = $9, bosques_demanda_semanal = $10, orden_inventario=$11 WHERE id_ingrediente = $6 RETURNING *';
     const result = await client.query(updateIngredientesQuery, [nombre, unidad, precio, proveedor, proveedor_id, id, store_route_order, producto_clave, moral_demanda_semanal, bosques_demanda_semanal, orden_inventario]);
 
-    const deleteFrecuenciasQuery = 'DELETE FROM ingredientes_frecuencia WHERE id_ingrediente = $1';
-    await client.query(deleteFrecuenciasQuery, [id]);
+    if (frecuencias_inventario) { // Check if frecuencias_inventario is not null
+      const deleteFrecuenciasQuery = 'DELETE FROM ingredientes_frecuencia WHERE id_ingrediente = $1';
+      await client.query(deleteFrecuenciasQuery, [id]);
 
-    const insertFrecuenciasQuery = 'INSERT INTO ingredientes_frecuencia (id_ingrediente, frecuencia_inventario_id) VALUES ($1, $2)';
-    for (let frecuencia of frecuencias_inventario) {
-      await client.query(insertFrecuenciasQuery, [id, frecuencia]);
+      const insertFrecuenciasQuery = 'INSERT INTO ingredientes_frecuencia (id_ingrediente, frecuencia_inventario_id) VALUES ($1, $2)';
+      for (let frecuencia of frecuencias_inventario) {
+        if (frecuencia) { // Check if frecuencia is not null
+          await client.query(insertFrecuenciasQuery, [id, frecuencia]);
+        }
+      }
     }
 
     await client.query('COMMIT');
